Amend PopoloObject to prepare for read/write data objects
In new system, rather than the properties of a PopoloObject
being defined as @property-ed functions accessing properties
of self.data - this will be replaced by the 'Attribute' class
 - that given a attribute will control getting and setting
 to self.data.

Subclasses of Attribute control different ways of talking to data
e.g. talking to data that is a date.

A small meta-class is used for PopoloObject to avoid repetition
in definitions.
e.g. person = Property(attr=person) can be expressed as
person = Property()